+++ server/api/pies.js
@@ -1,18 +1,8 @@
 const router = require('express').Router();
 const {models: { Pie, User }} = require('../db');
-
+const requireAdminToken = require('./gatekeeper')
 module.exports = router;
 
-const requireAdminToken = async (req, res, next) => {
-  try {
-    const token = req.headers.authorization;
-    const user = await User.findByToken(token);
-    if(user.type === 'admin') req.admin = user;
-    next();
-  } catch(error) {
-    next(error);
-  }
-};
 // GET /api/pies
 router.get('/', async (req, res, next) => {
   try {
